// shared-logic/model-review/model-review-v2.d.ts
export declare class ModelReviewV2 {
    id: string;
    idPosting: string;
    idHost: string;
    idUser: string;
    ratings: number;
    userFullName: string;
    photoUserUrl: string;
    comment: string;
    reply: any[];
    idAdmin: string;
    timestamp: any;
    constructor(id?: string, idPosting?: string, idHost?: string, idUser?: string, ratings?: number, userFullName?: string, photoUserUrl?: string, comment?: string, reply?: any[], idAdmin?: string, timestamp?: any);
}